requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
openai>=1.50.0
google-generativeai==0.7.2
requests==2.32.3
pydantic==2.9.2
python-slugify==8.0.4
pandas==2.2.3
numpy==2.1.2
